src/main/java/com/higginsn/amazon/linkedlists/MergeTwoSortedLists.java
package com.higginsn.amazon.linkedlists;

import com.higginsn.models.ListNode;

/**
 * https://leetcode.com/explore/interview/card/amazon/77/linked-list/2976/
 */
public class MergeTwoSortedLists {

    public ListNode mergeTwoLists(ListNode list1, ListNode list2) {
        if (list1 == null) {
            return list2;
        } else if (list2 == null) {
            return list1;
        }

        ListNode head;
        if (list1.val < list2.val) {
            head = list1;
            list1 = list1.next;
        } else {
            head = list2;
            list2 = list2.next;
        }

        ListNode tail = head;
        while (list1 != null && list2 != null) {
            if (list1.val < list2.val) {
                tail.next = list1;
                list1 = list1.next;
            } else {
                tail.next = list2;
                list2 = list2.next;
            }
            tail = tail.next;
        }

        if (list1 == null) {
            list1 = list2;
        }
        tail.next = list1;
        return head;
    }
}